project/requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
distribute==0.6.24
django-filebrowser==3.5.7
django-grappelli==2.5.6
sorl-thumbnail==12.1c
wsgiref==0.1.2
